+++ js/widgets/progressbar.js
@@ -1,4 +1,4 @@
-$.widget( "metro.progressBar" , {
+$.widget( "metro.progress" , {
 
     version: "3.0.0",
 
@@ -6,6 +6,7 @@ $.widget( "metro.progressBar" , {
         color: 'default',
         colors: false,
         value: 0,
+        animate: false,
         onProgress: function(value){}
     },
 
@@ -14,6 +15,10 @@ $.widget( "metro.progressBar" , {
     _create: function () {
         var that = this, element = this.element, o = this.options;
         var bar = element.children('.bar:last-child');
+
+        if (!element.hasClass('progress')) {
+            element.addClass('progress');
+        }
 
         $.each(element.data(), function(key, value){
             if (key in o) {
@@ -37,10 +42,10 @@ $.widget( "metro.progressBar" , {
             });
         }
 
-        this.progress(o.value);
+        this.set(o.value);
         this.color(o.color);
 
-        element.data('progressBar', this);
+        element.data('progress', this);
 
     },
 
@@ -62,7 +67,7 @@ $.widget( "metro.progressBar" , {
         o.color = value;
     },
 
-    progress: function(value){
+    set: function(value){
         if (value !== undefined) {
             var element = this.element, o = this.options, colors = this.colorsDim;
             var bar = element.children('.bar:last-child');
@@ -85,9 +90,26 @@ $.widget( "metro.progressBar" , {
 
             o.value = value;
 
-            bar.animate({
-                width: o.value + '%'
-            }, 100, function(){
+            if (o.animate !== false) {
+                var ani_speed = isNaN(o.animate) ? 500 : o.animate;
+                bar.animate({
+                    width: o.value + '%'
+                }, ani_speed, function(){
+                    if (typeof o.onProgress === 'function') {
+                        o.onProgress(value);
+                    } else {
+                        if (typeof window[o.onProgress] === 'function') {
+                            window[o.onProgress](value);
+                        } else {
+                            var result = eval("(function(){"+o.onProgress+"})");
+                            result.call(value);
+                        }
+                    }
+                });
+            } else {
+                bar.css({
+                    width: o.value + '%'
+                });
                 if (typeof o.onProgress === 'function') {
                     o.onProgress(value);
                 } else {
@@ -98,7 +120,8 @@ $.widget( "metro.progressBar" , {
                         result.call(value);
                     }
                 }
-            });
+            }
+
         } else {
             return this.options.value;
         }