+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.0.11
 * add keypad onChange event
+* chg progressBar animation model
+* fix box sizing for FF
+* chg progressBar role to progress and progress method to set
+* chg tabControl role to tabcontrol
 
 ## 3.0.10
 * fix modern input initialize with value